Fusion Systems Refactoring and debug display
